Remove typename for static fields
You need typename only for types. Moreover, it's actually the reason why we have to use it in this context. We use it to differentiate type from static field. That's why it's called typename.

But MSVC doesn't tell that it's actually an error. Quick search in google showed that MSVC late parsing or something that helps it identify what it is even without using typename.

It's actually good, but other compilers can't do that. Is there an option to make MSVC to treat these things as errors?